AsynchronousCSTest fixes
Include the CompletableFutureHelper in the test application

Separate the future that we complete from the one that the user's method
returns. Otherwise we set up a thenApply to get a new CompletionStage,
then we manually complete that new stage which means the thenApply never
runs. We need to complete the original CompletionStage instead.

Assert properly on exceptions thrown from future.get(). Need to assert
that an ExecutionException is thrown and that the cause is the exception
thrown from the user's method.

Avoid calling waitUntilCompleted after asserting that future.get()
throws an exception. We don't need to wait at this point because calling
get() will wait for it to finish, and the waitUntilCompleted method will
always fail because we know the future completed exceptionally.