Test review fixes
The test was refactored regarding the review comments now is more simple but
clear in what is the purpose of this test that is validate in which order
are executed the interceptors interacting with FT interceptors.

The test now is validating the call order of the interceptors
using a ConcurrentLinkedQueue from the concurrent java API
that uses FIFO queue strategy to store the data, this means that
the head of the queue is that element that has been
on the queue the longest time, so in this case the EarlyOrderFtInterceptor
that was called at first so is the HEAD of the queue and so on.

Signed-off-by: carlos de la rosa <kusanagi12002@gmail.com>