Don't reset the start time in testAsyncClassLevelTimeout
The test asserts that the execution time of the async method is correct,
therefore we should measure from when the method is launched, not from
when we start waiting for the result.

We already record the start time before we launch the method, so we just
need to remove the line where we were resetting the start time before we
start waiting for the result.

This is the same issue as #292

Signed-off-by: Andrew Rouse <anrouse@uk.ibm.com>