Reduce bounds on interceptor priority tests
For FaultToleranceInterceptorTest, the early and late interceptors were
far enough apart that they would be passed by a FT interceptor with any
priority between 1000 and 4099. I have reduced this to between 4001 and
4099.

For FaultToleranceInterceptorPriorityChangeAnnotationConfTest, the test
expected the interceptor to be earlier than the early interceptor,
meaning that it would be passed by a FT interceptor with a priority
below 3800. I have changed the test to require the FT interceptor to be
between 3800 and 3900.

Note that previously, both tests could be passed by an interceptor with
a static priority in the range 1000-3799.

Signed-off-by: Andrew Rouse <anrouse@uk.ibm.com>