clarify circuit breaker metrics
In case `@CircuitBreaker(failOn = {IllegalArgumentException.class})`
is defined and the guarded method throws e.g. `IllegalStateException`,
the circuit breaker treats it as success and therefore the
`circuitbreaker.callsSucceeded.total` metric needs to be incremented.

This commit adjusts the spec to make this clear, and also adds
a TCK test for this behavior.

Signed-off-by: Ladislav Thon <lthon@redhat.com>