Removed references to config from TimeoutTest
TimeoutTest had some code to configure timeouts, but it was all
inactive.

Allowing the user to configure timeouts here would somewhat undermine
what the test is trying to do (check that Timeout works with default
values, and when configured with larger and smaller values).

If we need to make the timeouts in this test configurable, we would have
to reconsider whether testing all of the scenarios in this test class is
necessary.

Signed-off-by: Andrew Rouse <anrouse@uk.ibm.com>