Remove strict timing check from testRestryWithNoDelayAndJitter
We're not trying to test performance and the strict timing check isn't
necessary here.

Also strengthen the check for `positiveDelays()` by discarding the first
delay which would be the time between object creation and first method
execution. (I'd _expect_ the RetryClientWithNoDelayAndJitter object to
be created just before the serviceA method is first called, but it's not
obvious to me that that _must_ be the case).

Signed-off-by: Andrew Rouse <anrouse@uk.ibm.com>