Pass the TCKConfig to the test application
In the case of a managed or remote arquillian container, a system
property set on the test process may not be set when a test application
runs.

We can't easily set system properties in all environments, so the
easiest way to pass the config properties is to include them in the test
application.

The TCKConfigArchiveAppender adds a jar with a resource containing the
configured multiplier and TCKConfig is updated to read this resource if
it is present.

I also consolidated our extension classes.

Signed-off-by: Andrew Rouse <anrouse@uk.ibm.com>