Test Historgrams using getValues rather than percentiles
MP Metrics Historgrams can be weighted so that more recent values have a
greater affect on the averages and percentiles than older ones.

Unfortunately, this means that changing the timeouts in the test can
change the results returned for averages and percentiles. When a longer
timeout is used, results added earlier in the test are older at the
point that the test takes a snapshot and so have a lower weighting when
compared to the same test run with a shorter timeout.

To avoid this unreliability, I have changed the tests to check the value
of Snapshot.getValues() which simply returns the values added to the
histogram. Since we're not trying to test the MP Metrics implementation,
it's sufficient to test that the correct values have been added to the
Histogram and not to test the computed averages.

Signed-off-by: Andrew Rouse <anrouse@uk.ibm.com>