move the note about naming/security context propagation to an extra section
This is because MicroProfile itself doesn't know anything about
naming or security context, and MicroProfile implementations
don't have to have anything like that. These terms only make sense
in a Java EE environment. Therefore, the requirement that naming
and security contexts are propagated should be present in an extra
section that only applies to Java EE container implementations.

Signed-off-by: Ladislav Thon <lthon@redhat.com>