MetricRegistry support on MP Metrics 4.0 or 5.0
MP Metrics 5.0 uses RegistryScope instead of RegistryType to inject the
MetricRegistry for a particular metrics scope.

Add a BaseRegistry qualifier and create a different provider for the v4
and v5 metrics APIs.

Some code needs to be compiled against the Metrics 5.0 API, so separate
that out into a separate module.

The common code is in the 5.0 module so that it doesn't have a
dependency on the rest of the TCK so that it can be built before the TCK
and its classes included in the TCK jar.

The version of MP Metrics API in use is detected based on the presence
of the RegistryScope annotation, or can be overridden by setting a
system property.