Avoid using test groups for test ordering (#649)
Test groups are global, not just for single test. This means that multiple
test classes may be "in progress" at the same time, which leads to issues
in at least one implementation which has global state.

Instead of test groups and `dependsOnGroups`, this commit uses simple
`dependsOnMethods`, which is local to the test class.